requirement.txt
openai
langchain_community
streamlit
transformers
python-dotenv